Move prisma calls to services
